feat(BP_character)
Creates the BP_StalkerAIPlayerCharacter Blueprint, which inherits from the C++ class SltakerAIPlayerCharacter. This asset will serve as the primary player character, allowing for visual adjustments and asset assignments in the editor.

Modifies the BP_ThirdPersonGameMode to set this new Blueprint as the Default Pawn Class. This ensures that the player possesses our custom character on level start, replacing the template default.